BTC-Bot/BTC-Bot/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.0.0",
    "requests>=2.25.0",
    "flask>=2.0.0"
]
